user_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
python-dotenv==1.0.0
redis==5.0.0
pydantic==2.4.2
httpx==0.25.0 